requirements.txt
requests
python-dotenv
redis
sqlite3